Add support for hashing arbitrary bytes (#21)
This modifies the existing functions for operating on individual set
members so, in addition to a key, they also accept a length. A new set
of functions with the suffix of "_str" have been added for users
operating on strings. These are thin wrappers around the base functions
that just call _strlen(3)_ for the user.