Revise type handling
Declare the set result types enums instead of "#define" states to make
it more idiomatic and, for the set_cmp function, to make it possible for
compilers to warn about unhandled conditions when using a "switch"
statement with the result. Since there is nothing preventing a set
member from having an index of 0, SET_TRUE has been revised to be -1
instead of 0 to avoid overlapping return values for functions that can
return counts or indexes in addition to status results. Similarly,
SET_ALREADY_PRESENT has been made a negative value for the same reason;
its previous value of 1 is also a valid index. The functions
__assign_node, set_clear and set_destroy always returned SET_TRUE, so
their return type has been changed to "void" and the return statements
deleted.